New version
Now there are 2 identical levels, but this is for testing transitions, menus, etc.
The music for the level is played like this: when you start the game level (from the main menu), one of the melodies is randomly selected, as soon as it ends, one of the two melodies is randomly selected again. When moving from level to level, the music does not change (does not turn off).